requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0
pillow==10.2.0
pytesseract==0.3.10
edge-tts==7.0.0
python-magic==0.4.27
opencv-python==4.9.0.80
numpy<2